Harden auth, input validation, and container security
- LoadSimulator.SimulateMemoryLoad / NodeBenchmark.RunMemoryBenchmark:
  cast to long before *1024*1024 to avoid int overflow on large inputs.
- LoadConfigApiController.Update: add a per-setting upper bound on
  Min/Max so a client can't push absurd/overflow-adjacent values into
  LoadConfig.
- NodeBenchmarkApiController.Run / AuthApiController.Login: replace
  int.Parse(config ?? "default") with int.TryParse + fallback, so a
  malformed config value degrades to the default instead of crashing
  the request.
- AuthApiController.Login: compare username/password with
  CryptographicOperations.FixedTimeEquals instead of `!=` to avoid a
  timing side-channel; validate Jwt:Key/Issuer/Audience explicitly
  (clear 500 instead of NRE via `!`); add rate limiting.
- Program.cs: register a per-IP fixed-window rate limiter ("login",
  5/min) and wire it into the pipeline, so /api/auth/login can't be
  brute-forced without throttling other clients.
- OptionalAuthorizationHandler: also require the "admin" role claim
  (which JwtTokenProvider already issues) instead of accepting any
  authenticated token.
- Dockerfile: run the final stage as the built-in non-root "app" user
  instead of root.

+++ ScaleTrigger/LoadSimulator.cs
@@ -52,7 +52,7 @@ namespace ScaleTrigger
                 return;
             }
 
-            int sizeInBytes = megabytes * 1024 * 1024;
+            long sizeInBytes = (long)megabytes * 1024 * 1024;
             byte[] buffer = new byte[sizeInBytes];
 
             Random.Shared.NextBytes(buffer);
